Add files via upload
As an accounting major, the importance of financial statement analysis is at the forefront of all things we consider when releasing financial statements. We need to be able to quickly analyze a company’s financial position in order to accurately release an opinion for ongoing business decisions. This project will help accountants and finance professionals quickly analyze the core financial ratios of a company.